Refactor: Restructured meme feature packages and moved MVI components
- Restructured meme feature packages for better organization.
- Moved MemeListScreen, MemeListTopAppBar, MemeListViewModel, and MemeListNavigation to the `screens.meme_list` package.
- Moved MVI components (MemeListState, MemeListAction) into MemeListViewModel.
- Created a root composable function MemeListScreenRoot for the meme list screen.
- Created a MemeCreation screen and view model.
- Updated NavigationRoot and NavGraph to reflect these changes.
- Renamed memeDataModule to dataModule.
- Created new folders `meme_creation`, `meme_list` and moved files.
- Moved `MasterMemeBackground` to `design_system` package.
