+++ app/src/main/java/com/mobilecampus/mastermeme/App.kt
@@ -1,9 +1,9 @@
 package com.mobilecampus.mastermeme
 
 import android.app.Application
-import com.mobilecampus.mastermeme.meme.data.di.memeDataModule
-import com.mobilecampus.mastermeme.meme.domain.di.domainModule
-import com.mobilecampus.mastermeme.meme.presentation.di.presentationModule
+import com.mobilecampus.mastermeme.meme.di.dataModule
+import com.mobilecampus.mastermeme.meme.di.domainModule
+import com.mobilecampus.mastermeme.meme.di.presentationModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.ext.koin.androidLogger
 import org.koin.core.context.GlobalContext.startKoin
@@ -15,7 +15,7 @@ class App : Application() {
         startKoin {
             androidLogger(Level.ERROR)
             androidContext(this@App)
-            modules(listOf(memeDataModule, domainModule, presentationModule))
+            modules(listOf(dataModule, domainModule, presentationModule))
         }
     }
 }
