+++ app/src/main/java/com/mobilecampus/mastermeme/core/presentation/design_system/MasterMemeBackground.kt
@@ -1,4 +1,4 @@
-package com.mobilecampus.mastermeme.core.presentation.ui
+package com.mobilecampus.mastermeme.core.presentation.design_system
 
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.LocalAbsoluteTonalElevation
