+++ app/src/main/java/com/mobilecampus/mastermeme/meme/di/DataModule.kt
@@ -1,4 +1,4 @@
-package com.mobilecampus.mastermeme.meme.data.di
+package com.mobilecampus.mastermeme.meme.di
 
 import androidx.room.Room
 import com.mobilecampus.mastermeme.meme.data.local.database.MemeDatabase
@@ -6,7 +6,7 @@ import com.mobilecampus.mastermeme.meme.data.local.datasource.MemeLocalDataSourc
 import com.mobilecampus.mastermeme.meme.domain.data_source.MemeDataSource
 import org.koin.dsl.module
 
-val memeDataModule = module {
+val dataModule = module {
     single { Room.databaseBuilder(get(), MemeDatabase::class.java, "meme_database").build() }
     single { get<MemeDatabase>().memeDao() }
     single<MemeDataSource> { MemeLocalDataSourceImpl(get(), get()) }
