+++ app/src/main/java/com/mobilecampus/mastermeme/meme/di/PresentationModule.kt
@@ -1,6 +1,6 @@
-package com.mobilecampus.mastermeme.meme.presentation.di
+package com.mobilecampus.mastermeme.meme.di
 
-import com.mobilecampus.mastermeme.meme.presentation.meme_list.MemeListViewModel
+import com.mobilecampus.mastermeme.meme.presentation.screens.meme_list.MemeListViewModel
 import org.koin.core.module.dsl.viewModelOf
 import org.koin.dsl.module
