--- app/src/main/java/com/mobilecampus/mastermeme/meme/presentation/screens/list/MemeListViewModel.kt
@@ -1,5 +1,0 @@
-package com.mobilecampus.mastermeme.meme.presentation.screens.list
-
-import androidx.lifecycle.ViewModel
-
-class MemeListViewModel : ViewModel(){}
