--- app/src/main/java/com/mobilecampus/mastermeme/meme/presentation/screens/list/mvi/MemeListAction.kt
@@ -1,5 +1,0 @@
-package com.mobilecampus.mastermeme.meme.presentation.screens.list.mvi
-
-sealed interface MemeListAction {
-    data object OnCreateMemeClick : MemeListAction
-}
