+++ app/src/main/java/com/mobilecampus/mastermeme/meme/presentation/screens/meme_creation/MemeCreation.kt
@@ -1,0 +1,9 @@
+package com.mobilecampus.mastermeme.meme.presentation.screens.meme_creation
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+fun MemeCreation(modifier: Modifier = Modifier) {
+    
+}
