+++ app/src/main/java/com/mobilecampus/mastermeme/meme/presentation/screens/meme_creation/MemeCreationViewModel.kt
@@ -1,0 +1,2 @@
+package com.mobilecampus.mastermeme.meme.presentation.screens.meme_creation
+
